landing page and auth

+++ Backend/controllers/expenseController.js
@@ -1,7 +1,7 @@
 import Expense from "../models/expenseModel.js"
 export const getAllExpense = async (req, res) => {
    try {
-      const expense = await Expense.find()
+      const expense = await Expense.find({ user: req.user.id })
       res.json({ success: true, count: expense.length, data: expense })
 
    } catch (error) {
@@ -17,7 +17,14 @@ export const createExpense = async (req, res) => {
       const { description, amount, category, date, notes } = req.body
       console.log(req.body);
       
-      const expense = new Expense({ description, amount, category, date, notes })
+      const expense = new Expense({ 
+         description, 
+         amount, 
+         category, 
+         date, 
+         notes,
+         user: req.user.id 
+      })
       const newExpense = await expense.save()
       res.status(201).json({ success: true, data: newExpense })
 
@@ -31,6 +38,17 @@ export const createExpense = async (req, res) => {
 export const updateExpense = async (req, res) => {
 
    try {
+      const expense = await Expense.findById(req.params.id)
+
+      if (!expense) {
+         return res.status(404).json({ success: false, message: "Expense Not Found" })
+      }
+
+      // Make sure the logged in user matches the expense user
+      if (expense.user.toString() !== req.user.id) {
+         return res.status(401).json({ success: false, message: "User not authorized" })
+      }
+
       const updateExpense = await Expense.findByIdAndUpdate(
          req.params.id,
          req.body,
@@ -52,16 +70,21 @@ export const updateExpense = async (req, res) => {
 
 export const deleteExpense = async (req, res) => {
    try {
-      const deleted = await Expense.findByIdAndDelete(req.params.id)
+      const expense = await Expense.findById(req.params.id)
+
+      if (!expense) {
+         return res.status(404).json({ success: false, message: "Not Found" })
+      }
+
+      // Check for user
+      if (expense.user.toString() !== req.user.id) {
+         return res.status(401).json({ success: false, message: "User not authorized" })
+      }
+
+      await Expense.findByIdAndDelete(req.params.id)
       res.json({ success: true, message: "Expense Deleted" })
 
-            console.log("Delete Function"+req.params.id);
-
-
-      if (!deleted) {
-         return res.status(404).json({ success: false, message: "Not Found" })
-
-      }
+      console.log("Delete Function"+req.params.id);
 
    } catch (error) {
       res.status(500).json({message: error.message })
